requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
importlib-metadata==3.4.0
isodate==0.6.0
lxml==4.6.2
Markdown==3.3.3
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zeep==4.0.0
zipp==3.4.0
